feat: split up script manager
