+++ packages/database/package.json
@@ -60,9 +60,9 @@
     "test:server": "jest -c ./tests/configs/server.ts --runInBand",
     "test:pull": "jest -c ./tests/configs/pull.ts --runInBand",
     "test:w": "jest -c ./tests/configs/w.ts --runInBand",
-    "script": "node ./scripts.js",
     "graph": "madge --image graph.png",
-    "default:env": "sh ./bash/setup.sh",
+    "script": "node ./scripts/compile-package-scripts.js",
+    "default:env": "sh ./scripts/setup.sh",
     "module:get-codes": "ts-node ./src/api/module/get-codes.ts",
     "module:get": "ts-node ./src/api/module/get.ts",
     "module:pull": "ts-node ./src/api/module/pull.ts",
