+++ packages/database/scripts/setup.sh
